requirements.txt
fastapi==0.115.12
python-dotenv==1.1.0
botbuilder-core==4.16.2
botbuilder-schema==4.16.2
botframework-connector==4.16.2
uvicorn==0.34.0
agno==1.2.13
groq==0.20.0
lancedb==0.21.2
fastembed==0.6.0
pydantic==2.11.1
requests==2.32.3
msal==1.32.0
PyJWT==2.10.1
aiohttp==3.11.15
numpy==2.2.4
pyarrow==19.0.1
pandas==2.2.3